+++ src/main/java/com/example/carsharing/service/UserDbService.java
@@ -1,7 +1,6 @@
 package com.example.carsharing.service;
 
 import com.example.carsharing.domain.User;
-import com.example.carsharing.domain.UserDto;
 import com.example.carsharing.error.user.UserNotFoundException;
 import com.example.carsharing.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -19,16 +18,22 @@ public class UserDbService {
         return userRepository.findAll();
     }
 
-    public User getUserById(Long id){
-        return userRepository.findById(id).orElseThrow(()->new UserNotFoundException(id));
+    public User getUserById(Long id) {
+        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
     }
-    public User saveUser(User user){
+
+    public User saveUser(User user) {
         return userRepository.save(user);
     }
-    public void deleteUser(Long id){
+
+    public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
     }
-    public User updateUser(Long id, User updatedUser){
+
+    public User updateUser(Long id, User updatedUser) {
         User existingUser = getUserById(id);
 
         existingUser.setName(updatedUser.getName());
